+++ helpers/constants.ts
@@ -14,6 +14,7 @@ export const ECOSYSTEM_VESTING_BENEFICIARY = DAO_TREASURY_ADDRESS;
 export const REALM_DIAMOND = "0x1D0360BaC7299C86Ec8E99d0c1C9A95FEfaF2a11";
 export const GHST_ADDRESS = "0x385Eeac5cB85A38A9a07A70c73e0a3271CfB54A7";
 export const QUICKSWAP_ROUTER_ADDRESS = "0xa5E0829CaCEd8fFDD4De3c43696c57F7D7A678ff";
+export const VESTING_START_TIME = BigNumber.from(1647759600); // March 20, 2022 @ 12:00am (UTC)
 
 export const PROXY_ADMIN_ADDRESS = "0xbE50b0594edeA37C34dF287778d85B9546Fdb309";
