+++ helpers/helpers.ts
@@ -21,6 +21,14 @@ import {
   GWEI,
   WMATIC_ADDRESS,
   PERMIT_TYPES,
+  FUD_PARAMS,
+  FOMO_PARAMS,
+  ALPHA_PARAMS,
+  KEK_PARAMS,
+  INITIAL_ALCHEMICA_SEED,
+  QUICKSWAP_ROUTER_ADDRESS,
+  GHST_ADDRESS,
+  VESTING_START_TIME,
 } from "./constants";
 
 import {
@@ -35,6 +43,12 @@ import {
   address,
   sleep,
 } from "./utils";
+
+import {
+  AlchemicaToken,
+  AlchemicaVesting,
+  IUniswapV2Router02,
+} from "../typechain"
 
 export async function createDomain(
   token: Contract,
@@ -228,5 +242,129 @@ export async function verify(verifyParams: VerifyParams[]): Promise<void> {
       console.log("Contract verification failed." )
       console.log(e);
     }
+  }
+}
+
+export async function deployVestingContracts(
+  owner: Signer,
+  proxyAdmin: Contract,
+) {
+  let returnParams: VerifyParams[] = [];
+  let vestingImplementation = await deployVestingImplementation(owner);
+  await sleep(10000);
+  console.log("Vesting Implementation: " + vestingImplementation.contract.address)
+  returnParams.push(vestingImplementation);
+  await vestingImplementation.contract.deployed();
+
+  let ecosystemVestingProxy = await deployAndInitializeVestingProxy(
+    owner,
+    vestingImplementation.contract,
+    await address(owner),
+    proxyAdmin,
+    VESTING_START_TIME,
+    ETHER.div(10), // 10% decay per year
+    true,
+  );
+  console.log("Ecosystem Vesting: " + ecosystemVestingProxy.contract.address);
+  returnParams.push(ecosystemVestingProxy);
+  let gameplayVestingProxy = await deployAndInitializeVestingProxy(
+    owner,
+    vestingImplementation.contract,
+    await address(owner),
+    proxyAdmin,
+    VESTING_START_TIME,
+    ETHER.div(10), // 10% decay per year
+    true,
+  )
+  console.log("Gameplay Vesting: " + gameplayVestingProxy.contract.address);
+  returnParams.push(gameplayVestingProxy);
+  return returnParams;
+}
+
+export async function deployAlchemica(
+  owner: Signer,
+  proxyAdmin: Contract,
+  realmDiamond: string,
+  gameplayVestingContract: Contract,
+  ecosystemVestingContract: Contract,
+) {
+  let returnParams: VerifyParams[] = [];
+  let alchemicaImplementation = await deployAlchemicaImplementation(owner);
+  await sleep(10000);
+  console.log("Alchemica Implementation: " + alchemicaImplementation.contract.address);
+  returnParams.push(alchemicaImplementation);
+  await alchemicaImplementation.contract.deployed();
+
+  for(let params of [FUD_PARAMS, FOMO_PARAMS, ALPHA_PARAMS, KEK_PARAMS]) {
+    let alchemicaProxy = await deployAndInitializeAlchemicaProxy(
+      owner,
+      alchemicaImplementation.contract,
+      proxyAdmin,
+      params.name,
+      params.symbol,
+      params.supply,
+      realmDiamond,
+      gameplayVestingContract,
+      ecosystemVestingContract,
+
+    );
+    console.log(params.name + ": " + alchemicaProxy.contract.address);
+    returnParams.push(alchemicaProxy);
+  }
+
+  return returnParams;
+}
+
+export async function releaseAndLP(
+  beneficiary: Signer,
+  vestingContract: AlchemicaVesting,
+  alchemicas: [AlchemicaToken, AlchemicaToken, AlchemicaToken, AlchemicaToken],) 
+{
+  const alchIn = INITIAL_ALCHEMICA_SEED[0];
+  const ghstIn = INITIAL_ALCHEMICA_SEED[1];
+  const names = ["FUD", "FOMO", "ALPHA", "KEK"];
+  const router: IUniswapV2Router02 = (await hre.ethers.getContractAt("IUniswapV2Router02", QUICKSWAP_ROUTER_ADDRESS)) as IUniswapV2Router02;
+  const ghst = await hre.ethers.getContractAt("contracts/interfaces/IERC20.sol:IERC20", GHST_ADDRESS);
+  let tx = await vestingContract.connect(beneficiary).batchRelease(
+    [
+      await address(alchemicas[0]),
+      await address(alchemicas[1]),
+      await address(alchemicas[2]),
+      await address(alchemicas[3]),
+    ],
+  );
+  await tx.wait();
+  tx = await ghst.connect(beneficiary).approve(
+    await address(router),
+    ghstIn[0].add(ghstIn[1]).add(ghstIn[2]).add(ghstIn[3]),
+  );
+  await tx.wait();
+  console.log("GHST approved");
+  for(let i = 0; i < alchemicas.length; i++) {
+    let alchemica = alchemicas[i];
+    console.log(
+      (await alchemica.balanceOf(await address(beneficiary))).toString() + 
+      " " +  names[i] + " received."
+    );
+    tx = await alchemica.connect(beneficiary).approve(await address(router), alchIn[i]);
+    await tx.wait();
+    console.log(names[i] + " approved");
+
+    await sleep(10000);
+    tx = await router.connect(beneficiary).addLiquidity(
+      await address(alchemica),
+      GHST_ADDRESS,
+      alchIn[i],
+      ghstIn[i],
+      0,
+      0,
+      await address(beneficiary),
+      await currentTimestamp() + 1000,
+    );
+    await tx.wait();
+    console.log(
+      alchIn[i].toString() + 
+      " " + names[i] + " (wei) added to liquidity pool with " + 
+      ghstIn[i].toString() + " GHST (wei)");
   }
 }
